+++ kamehouse-cmd/src/main/java/com/nicobrest/kamehouse/cmd/model/Operation.java
@@ -6,5 +6,6 @@ package com.nicobrest.kamehouse.cmd.model;
  * @author nbrest
  */
 public enum Operation {
+  DECRYPT,
   ENCRYPT
 }