+++ kamehouse-cmd/src/test/java/com/nicobrest/kamehouse/cmd/config/MainAppTest.java
@@ -30,11 +30,25 @@ public class MainAppTest {
   }
 
   /**
-   * Tests that the process executes correctly.
+   * Tests running kamehouse-cmd with operation decrypt.
+   */
+  @Test
+  public void decryptSuccessfulTest() throws IOException {
+    String[] args = new String[] { "-o", "decrypt", "-if", "in.enc", "-of", "out.dec"};
+    MainApp.main(args);
+    // no exceptions thrown
+
+    when(FileUtils.readFileToByteArray(any())).thenCallRealMethod();
+    MainApp.main(args);
+    // no exceptions thrown
+  }
+
+  /**
+   * Tests running kamehouse-cmd with operation encrypt.
    */
   @Test
   public void encryptSuccessfulTest() throws IOException {
-    String[] args = new String[] { "-o", "encrypt", "-if", "in.txt", "-of", "out.enc"};
+    String[] args = new String[] { "-o", "encrypt", "-if", "in.txt", "-of", "out.enc", "-v"};
     MainApp.main(args);
     // no exceptions thrown