+++ kamehouse-cmd/src/test/java/com/nicobrest/kamehouse/cmd/model/CmdArgumentHandlerTest.java
@@ -64,7 +64,6 @@ public class CmdArgumentHandlerTest {
     assertNull(cmdArgumentHandler.getArgument("of"));
   }
 
-
   /**
    * Tests an invalid operation.
    */